Renames Vector to OVector and NVector to Vector
