+++ test/junit/scala/collection/IndexedSeqTest.scala
@@ -584,11 +584,11 @@ package IndexedTestImpl {
     }
   }
 
-  import scala.collection.immutable.NVector
-  class NVectorTest extends ImmutableIndexedSeqTest[NVector[String], String]  with StringTestData {
+  import scala.collection.immutable.OVector
+  class OVectorTest extends ImmutableIndexedSeqTest[OVector[String], String]  with StringTestData {
 
-    override protected def underTest(size: Int): NVector[String] = {
-      val res = NVector.newBuilder[String]
+    override protected def underTest(size: Int): OVector[String] = {
+      val res = OVector.newBuilder[String]
       for (i <- 0 until size)
         res += expectedValueAtIndex(i)
       res.result()
